refactor: Change Dir Name (#172)
- req -> request
- res -> response
